fix: check `process.cwd` before calling it
Some runtimes (like React Native) has a `process` global, but doesn't provide a `cwd` function. Instead of assuming the `cwd` function is available, just because the `process` global is, merging this PR will update the check to actually verify the `process.cwd` is a callable function before relying on it.