fix: allow format calls with one side of (file + ext) missing
In node, the following are valid calls of `format`:

```ts
format({ file: 'foo' });
// foo

format({ dir: '/foo', name: 'bar' });
// /foo/bar

format({ dir: '/foo', ext: 'bar' });
// /foo/.bar
```

The last one is a little nonsensical, but the other two are valid
(extensionless files for example). We should also be matching node's
behaviour of course.

Also enables `strict: true` in TypeScript, which is what caught this.